requirements.txt
Django==4.1.3
'pyHanko[pkcs11,image-support,opentype,xmp]'
djangorestframework==3.14.0